+++ packages/components/src/navigationOption/NavigationOption.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import Types from 'prop-types';
 import requiredIf from 'react-required-if';
-
+import classNames from 'classnames';
 import Chevron from '../chevron';
 
 import Option from '../common/Option';
@@ -19,6 +19,7 @@ const NavigationOption = ({
   disabled,
   showMediaAtAllSizes,
   showMediaCircle,
+  className,
 }) => {
   const sharedProps = {
     as,
@@ -30,6 +31,7 @@ const NavigationOption = ({
     disabled,
     showMediaAtAllSizes,
     showMediaCircle,
+    className: classNames('tw-navigation-option', className),
   };
 
   if (href) {
@@ -38,7 +40,6 @@ const NavigationOption = ({
 
   return (
     <Option
-      className="tw-navigation-option"
       {...sharedProps}
       onClick={(event) => {
         if (!disabled) {
@@ -61,6 +62,7 @@ NavigationOption.propTypes = {
   disabled: Types.bool,
   showMediaAtAllSizes: Types.bool,
   showMediaCircle: Types.bool,
+  className: Types.string,
 };
 
 NavigationOption.defaultProps = {
@@ -73,6 +75,7 @@ NavigationOption.defaultProps = {
   disabled: false,
   showMediaAtAllSizes: false,
   showMediaCircle: true,
+  className: null,
 };
 
 export default NavigationOption;