+++ packages/components/src/typeahead/Typeahead.story.js
@@ -56,7 +56,7 @@ export const basic = () => {
       label: 'Globe',
     },
     {
-      label: 'British Pound',
+      label: 'British pound',
     },
     {
       label: 'Euro',