+++ packages/docs/liveEditorCode/typeahead.code.js
@@ -15,7 +15,7 @@
       label: 'Globe',
     },
     {
-      label: 'British Pound',
+      label: 'British pound',
     },
     {
       label: 'Euro',
@@ -29,7 +29,7 @@
   const allowNew = true;
   const validateChip =
     multiple && allowNew
-      ? option =>
+      ? (option) =>
           /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/.test(
             option.label,
           )